Make connection and keep-alive timeouts configurable
When a Nuxeo server is behind a load balancer which can reuse backend TCP
connections, we should ensure that the load balancer is responsible for closing
TCP connections by defining its idle timeout to be lower than Nuxeo's
keep-alive timeout [1].

However, as of today, Nuxeo's embedded Tomcat keep-alive timeout is implicitly
set to 20 seconds and cannot be easily configured as:
* In `server.xml.nxftl`, `connectionTimeout` is hard-coded to 20 seconds [2];
* And as `keepAliveTimeout`, the HTTP Connector attribute which defines this
keep-alive timeout, uses the value set for `connectionTimeout` as its default
value [3].

This 20 seconds value can be problematic in cases where the load balancer idle
timeout is higher and cannot be configured (for example, the default idle
timeout on AWS Application Load Balancers is 60s [4]).

In these cases, it would be useful to be able to configure this keep-alive
timeout.

This can already be done today by manually altering the contents of the
`server.xml.nxftl` template file but this would be better if the
`keepAliveTimeout` value could be parameterized.
The `connectionTimeout` value could also be parameterized on the same
occasion.

To do so, this commit adds `nuxeo.server.http.connectionTimeout` and
`nuxeo.server.http.keepAliveTimeout` configuration parameters, which can be
defined in `nuxeo.conf`.

See:
[1] https://doc.nuxeo.com/nxdoc/http-and-https-reverse-proxy-configuration/#configuration-between-aws-elastic-load-balancing-elb-and-apache
[2] https://github.com/nuxeo/nuxeo/blob/v11.1.17/server/nuxeo-nxr-server/src/main/resources/templates/common-base/conf/server.xml.nxftl#L84
[3] https://tomcat.apache.org/tomcat-9.0-doc/config/http.html#Standard_Implementation
[4] https://docs.aws.amazon.com/elasticloadbalancing/latest/application/application-load-balancers.html#connection-idle-timeout